Refactor pagination context deprecations
